requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.1
PyMySQL>=1.1.1
requests>=2.32.3
scikit-learn>=1.5.0
joblib>=1.4.2
numpy>=1.26.0
pandas>=2.2.0

